Redesigned the browse pages to *hopefully* be quicker. Fixes #11
The slow-down was primarily due to loads of spurious can_view and
can_edit calls to resources that are part of a collection. Literally
hundreds of DB queries. I tried to adjust the
ResourceCollection.jsonClosure function to be more efficient, even if
the code is a little sloppier. Also involved changing the interface to
do multiple AJAX requests when collections are expanded, rather than
loading the whole thing in through a django template.